third_party/WebKit/LayoutTests/inspector/sources/debugger/source-frame-inline-breakpoint-decorations-expected.txt

Checks that JavaScriptSourceFrame show inline breakpoints correctly


Running: testAddRemoveBreakpoint
Setting breakpoint
breakpoint at 3
  inline breakpoint at (3, 12)
  inline breakpoint at (3, 20) disabled
  inline breakpoint at (3, 30) disabled
  inline breakpoint at (3, 49) disabled
  inline breakpoint at (3, 56) disabled
Toggle breakpoint

Running: testAddRemoveBreakpointInLineWithOneLocation
Setting breakpoint
breakpoint at 5
  inline breakpoint at (5, 4)
  inline breakpoint at (5, 13) disabled
Toggle breakpoint

Running: clickByInlineBreakpoint
Setting breakpoint
breakpoint at 3
  inline breakpoint at (3, 12)
  inline breakpoint at (3, 20) disabled
  inline breakpoint at (3, 30) disabled
  inline breakpoint at (3, 49) disabled
  inline breakpoint at (3, 56) disabled
Click by second breakpoint
breakpoint at 3
  inline breakpoint at (3, 12)
  inline breakpoint at (3, 20)
  inline breakpoint at (3, 30) disabled
  inline breakpoint at (3, 49) disabled
  inline breakpoint at (3, 56) disabled
Click by first breakpoint
breakpoint at 3
  inline breakpoint at (3, 12) disabled
  inline breakpoint at (3, 20)
  inline breakpoint at (3, 30) disabled
  inline breakpoint at (3, 49) disabled
  inline breakpoint at (3, 56) disabled
Click by second breakpoint

Running: toggleBreakpointInAnotherLineWontRemoveExisting
Setting breakpoint in line 4
breakpoint at 4
  inline breakpoint at (4, 9)
  inline breakpoint at (4, 28) disabled
  inline breakpoint at (4, 36) disabled
Setting breakpoint in line 3
breakpoint at 3
  inline breakpoint at (3, 12)
  inline breakpoint at (3, 20) disabled
  inline breakpoint at (3, 30) disabled
  inline breakpoint at (3, 49) disabled
  inline breakpoint at (3, 56) disabled
breakpoint at 4
  inline breakpoint at (4, 9)
  inline breakpoint at (4, 28) disabled
  inline breakpoint at (4, 36) disabled
Click by first inline breakpoints
